requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cleanup==5.1.0
django-environ==0.4.5
Pillow==8.0.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
